alexa-skill/pyproject.toml
[tool.poetry]
name = "cleaning-management-system"
version = "1.0.0"
description = "音声ベース掃除記録システム - Alexa Skills Kit + Google Sheets"
authors = ["Cleaning Management System <admin@example.com>"]
readme = "README.md"
packages = [{include = "src"}]
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
ask-sdk-core = "^1.19.0"
gspread = "^6.0.0"
google-auth = "^2.24.0"
google-auth-oauthlib = "^1.1.0"
requests-oauthlib = "^1.3.1"
oauthlib = "^3.2.2"
pyyaml = "^6.0.1"
boto3 = "^1.34.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.11.0"
flake8 = "^6.1.0"

# ASK CLI管理用スクリプト
[tool.poetry.scripts]
setup-ask-cli = "scripts.setup_ask_cli:main"
deploy-skill = "scripts.deploy_skill:main"
update-model = "scripts.update_model:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true 